requirements-docs.txt
colorama
fuzzywuzzy
pydantic
python-dotenv
python-Levenshtein
simplejson
sphinx
sphinx-press-theme